Updated to reflect PR review
