+++ src/Aeon.Acquisition/IsRepositoryClean.cs
@@ -6,14 +6,11 @@ using LibGit2Sharp;
 
 namespace Aeon.Acquisition
 {
-
     [Combinator]
-    [Description("From a Repository object determines if the repository is dirty or uncommited changes exist")]
+    [Description("Determines whether a specified repository is clean or if uncommitted or untracked changes exist")]
     [WorkflowElementCategory(ElementCategory.Transform)]
-
-    public class CheckDirtyRepo
+    public class IsRepositoryClean
     {
-
         public IObservable<Boolean> Process(IObservable<IRepository> source)
         {
             return source.Select(value =>
@@ -21,11 +18,8 @@ namespace Aeon.Acquisition
                 RepositoryStatus status = value.RetrieveStatus();
                 bool isDirtyRepostatus = status.IsDirty;
                 bool isUncommitedFiles = value.Diff.Compare<TreeChanges>().Count > 0;
-                return isDirtyRepostatus & isUncommitedFiles;
+                return !(isDirtyRepostatus | isUncommitedFiles);
             });
         }
-
-
     }
-
 }
